requirements.txt
langgraph
langchain
langchain-community
python-dotenv
fastapi
motor
langchain-google-genai
langchain-openai
langchain-pinecone
PyMuPDF
pytz
loguru
uvicorn[standard]
python-multipart
langchain_experimental
duckduckgo-search
pydantic[email]
python_jose==3.3.0
pillow
langchain-text-splitters
python-docx
cloudinary
PyJWT==2.8.0
pymupdf 
docx2txt
